Create To-Do-List
Develop a basic to-do app that enables a user to add daily tasks. The to-do web app must allow users to add new items to a list, and items added to a list must be visible. Also, a user can mark a task "complete" upon its successful completion, and these completed tasks must appear on a separate list, say Completed Tasks.

To further increase the web app's complexity or make it more detailed, you can add the date and time when a task is added and/or completed